backend/requirements.txt
alembic==1.14.0
fastapi==0.115.6
httpx==0.28.1
psycopg[binary]==3.2.3
pytest==8.3.4
python-dotenv==1.0.1
SQLAlchemy==2.0.36
uvicorn[standard]==0.32.1
